bot3/requirements.txt
# Telegram Bot 依赖
python-telegram-bot>=20.0,<21.0

# 数据库
aiosqlite>=0.19.0

# Excel 处理
openpyxl>=3.1.0

# 日期时间处理
pytz>=2023.3

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 任务调度
APScheduler>=3.10.0

# HTTP 请求（python-telegram-bot 依赖）
httpx>=0.24.0

# 其他工具
python-dateutil>=2.8.2

